refactor: [박태준] studyroom service 계층 인터페이스와 구현체 분리 [JSAB-60]
StudyRoomService를 StudyRoomServiceImpl 구현체로 변경
Controller가 인터페이스에 의존하도록 StudyRoomService interface 생성
