+++ src/main/java/com/wowraid/jobspoon/studyroom/service/StudyRoomService.java
@@ -2,77 +2,24 @@ package com.wowraid.jobspoon.studyroom.service;
 
 import com.wowraid.jobspoon.studyroom.controller.request_Form.RegisterStudyRoomRequestForm;
 import com.wowraid.jobspoon.studyroom.controller.request_Form.UpdateStudyRoomRequestForm;
-import com.wowraid.jobspoon.studyroom.entity.StudyRoom;
-import com.wowraid.jobspoon.studyroom.repository.StudyRoomRepository;
 import com.wowraid.jobspoon.studyroom.service.response.RegisterStudyRoomResponse;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-@RequiredArgsConstructor
-public class StudyRoomService {
+public interface StudyRoomService {
 
-    private final StudyRoomRepository studyRoomRepository;
+    // 스터디룸 생성
+    RegisterStudyRoomResponse createStudyRoom(RegisterStudyRoomRequestForm requestForm);
 
-    // 이건 생성이에옹
-    @Transactional
-    public RegisterStudyRoomResponse createStudyRoom(RegisterStudyRoomRequestForm requestForm){
-        StudyRoom newStudyRoom = StudyRoom.create(
-            requestForm.getStudyTitle(),
-            requestForm.getDescription(),
-            requestForm.getMaxMembers(),
-            requestForm.getStatus(),
-            requestForm.getRegion(),
-            requestForm.getChatLink()
-);
-        StudyRoom savedStudyRoom = studyRoomRepository.save(newStudyRoom);
-        return RegisterStudyRoomResponse.from(savedStudyRoom);
-    }
+    // 스터디룸 전체조회
+    List<RegisterStudyRoomResponse> findAllStudyRooms();
 
-    // 이건 전체조회에옹
-    @Transactional
-    public List<RegisterStudyRoomResponse> findAllStudyRooms(){
-        List<StudyRoom> studyRooms = studyRoomRepository.findAll();
+    // 스터디룸 지역별 조회
+    List<RegisterStudyRoomResponse> findStudyRoomsByRegion(String region);
 
-        return studyRooms.stream()
-                .map(RegisterStudyRoomResponse::from)
-                .collect(Collectors.toList());
-    }
+    // 스터디룸 수정
+    RegisterStudyRoomResponse updateStudyRoom(Long studyRoomId, UpdateStudyRoomRequestForm requestForm);
 
-    // 이건 필터조회에옹
-    @Transactional
-    public List<RegisterStudyRoomResponse> findStudyRoomsByRegion(String region){
-        return studyRoomRepository.findByRegion(region).stream()
-                .map(RegisterStudyRoomResponse::from)
-                .collect(Collectors.toList());
-    }
-
-    // 이건 수정이다람쥐?
-    @Transactional
-    public RegisterStudyRoomResponse updateStudyRoom(Long studyRoomId, UpdateStudyRoomRequestForm requestForm){
-        StudyRoom studyRoom = studyRoomRepository.findById(studyRoomId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 스터디룸을 찾을 수 없습니다."));
-
-        studyRoom.update(
-                requestForm.getStudyTitle(),
-                requestForm.getDescription(),
-                requestForm.getMaxMembers(),
-                requestForm.getStatus(),
-                requestForm.getRegion(),
-                requestForm.getChatLink()
-        );
-        return RegisterStudyRoomResponse.from(studyRoom);
-    }
-
-    // 한번씩~~ 너를 지울때마다~~ 가슴이 아려와~~(이승기:삭제)
-    @Transactional
-    public void deleteStudyRoom(Long studyRoomId){ // 추후 Account currentUser 추가해야하리보
-        StudyRoom studyRoom = studyRoomRepository.findById(studyRoomId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 스터디를 찾을 수 없습니다. id=" + studyRoomId));
-        studyRoomRepository.delete(studyRoom);
-    }
+    // 스터디룸 삭제
+    void deleteStudyRoom(Long studyRoomId);
 }
